Introduce partitions for table_log log table.
The basic idea is to establish a partitioned log table
that allows log table maintenance in a round robin manner:
If the DBA decides to archive and the clean a log table, he
sets the active log partition id to another one, backup the log
table data and use a TRUNCATE to clean it. He still needs to wait
unless all transactions holding a lock on this table will disappear,
however, TRUNCATE doesn't block remaining log actions on a specific
table.

The active partition is globally controlled by a custom GUC called
table_log.active_partition, which defaults to '1'. Currently two
partitions are possible. The log tables will have a _0 or _1 appended
to their identifiers. When partitions is planned to use, the DBA has
to call table_log_init() with all required parameters, plus all optional
ones, since the partition_mode parameter was added to the call site
with 5 arguments only and defaults to 'SINGLE'. This causes table_log
to still use the old behavior (for backward compatibility).

There's also a special view with a UNION ALL over the partitioned
log table of a relation, which is also used by table_log_restore_table()
to restore all requested data.

While at it, repair at least the handling of the restore table in
table_log_restore_table() which wasn't schema-aware. This lead to
choosing a random table (however OID comes first...) to restore to and
thus to random failures under certain circumstances. To fix this,
allow a schema-qualified restore table name which is then parsed
into its identifier components (schema.tablename). I really like to
change this SPI handling overall by using SysCache calls, but this means
a complete overhaul of the code.