Properly escape single quotes in do_quote_literal().
This is a hot fix to properly handle single quotes in SQL statements
used to modify the table_log log tables.

We need to do more about it, since it seems this ancient code isn't
well prepared for multibyte or other corner cases. The best way
to that in the future seems to be to rely on quote_literal() in the
backend, but this involves additional palloc()/memcpy() calls.
The concern is that this would have some impact to performance.

So commit this hot fix first and make the other solution happen after
some more performance tests.